+++ src/tensors/cpu/expression_graph_packable.h
@@ -207,6 +207,9 @@ public:
           Transpose10(tmp, val);
         }
         if (gemmElementType == Type::intgemm8) {
+#if defined(WASM)
+          ABORT("Int8::PrepareA is not implemented for wasm.");
+#else
           float quantMult = 127.0f / intgemm::MaxAbsolute(val->data(), val->data() + val->shape().elements());
           intgemm::Int8::PrepareA(tmp->data(), /*input*/
                                 paramMat->data<int8_t>(), /*output*/
@@ -215,7 +218,11 @@ public:
                                 cols(val));
           //Put the quantMult at the back of the tensor
           *(reinterpret_cast<float *>(paramMat->data<int8_t>() + val->shape().elements())) = quantMult;
+#endif
         } else {
+#if defined(WASM)
+          ABORT("Int16::PrepareA is not implemented for wasm.");
+#else
           float quantMult = 1024.0f;
           intgemm::Int16::PrepareA(tmp->data(), /*input*/
                                 paramMat->data<int16_t>(), /*output*/
@@ -224,6 +231,7 @@ public:
                                 cols(val));
           //Put the quantMult at the back of the tensor
           *(reinterpret_cast<float *>(paramMat->data<int16_t>() + val->shape().elements())) = quantMult;
+#endif
         }
 
         //Save... Same as the fbgemm case